pyproject.toml
[project]
name = "lightgrph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.124.2",
    "lightrag-hku>=1.4.9.8",
    "numpy>=2.3.5",
    "openai==0.28.1",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
